+++ sql_to_fnc.ts
@@ -3,8 +3,8 @@ import * as path from 'path';
 import { generateSQL } from './src/sql_to_fnc-sql.lib';
 import { generateAPI } from './src/sql_to_fnc-api.lib';
 import { FieldDefinition } from './src/sql_to_fnc.interfaces';
-import {generateTestE2E} from "./src/sql_to_fnc-test.lib";
-import {generateAngularModule} from "./src/sql_to_fnc-www.lib";
+import { generateTestE2E } from './src/sql_to_fnc-test.lib';
+import { generateAngularModule } from './src/sql_to_fnc-www.lib';
 
 const args = process.argv.slice(2);
 if (!args[0]) {